+++ .eslintrc.json
@@ -4,7 +4,7 @@
     "es2021": true,
     "jest/globals": true
   },
-  "extends": ["plugin:react/recommended", "standard"],
+  "extends": ["plugin:react/recommended", "standard", "prettier"],
   "parser": "@typescript-eslint/parser",
   "parserOptions": {
     "ecmaFeatures": {
@@ -13,7 +13,13 @@
     "ecmaVersion": "latest",
     "sourceType": "module"
   },
-  "plugins": ["react", "react-native", "@typescript-eslint", "jest"],
+  "plugins": [
+    "react",
+    "react-native",
+    "@typescript-eslint",
+    "jest",
+    "prettier"
+  ],
   "settings": {
     "react": {
       "version": "detect"
