+++ lib/dependencies/DependencyValues.ts
@@ -1,4 +1,4 @@
-import crypto from "crypto";
+import { DependencyKey } from "./DependencyKey";
 
 export interface ImmutableDependencyValues {
   get: <T>(key: DependencyKey<T>) => T;
@@ -30,28 +30,3 @@ export class DependencyValues implements ImmutableDependencyValues {
     return newValues;
   }
 }
-
-type DependencyKey<T> = {
-  readonly identifier: string;
-  readonly createDefaultValue?: (values: ImmutableDependencyValues) => T;
-};
-
-export const createDependencyKey = <T>(
-  createDefaultValue?: ((values: ImmutableDependencyValues) => T) | T
-): DependencyKey<T> => ({
-  identifier: crypto.randomUUID(),
-  createDefaultValue: !!createDefaultValue
-    ? _makeDefaultValueCreation(createDefaultValue)
-    : undefined,
-});
-
-const _makeDefaultValueCreation = <T>(
-  create: ((values: ImmutableDependencyValues) => T) | T
-) => {
-  return (values: ImmutableDependencyValues) => {
-    if (create instanceof Function) {
-      return create(values);
-    }
-    return create;
-  };
-};
