+++ lib/dependencies/index.ts
@@ -1,0 +1,5 @@
+export { DependencyKey, createDependencyKey } from "./DependencyKey";
+export {
+  DependencyValues,
+  ImmutableDependencyValues,
+} from "./DependencyValues";
