+++ package.json
@@ -78,10 +78,12 @@
     "@typescript-eslint/parser": "^5.51.0",
     "babel-preset-expo": "~9.1.0",
     "eslint": "^8.33.0",
+    "eslint-config-prettier": "^8.6.0",
     "eslint-config-standard": "^17.0.0",
     "eslint-plugin-import": "^2.27.5",
     "eslint-plugin-jest": "^27.2.1",
     "eslint-plugin-n": "^15.6.1",
+    "eslint-plugin-prettier": "^4.2.1",
     "eslint-plugin-promise": "^6.1.1",
     "eslint-plugin-react": "^7.32.2",
     "eslint-plugin-react-native": "^4.0.0",
@@ -89,6 +91,7 @@
     "jest": "^26.6.3",
     "jest-expo": "^45.0.0",
     "lint-staged": "^13.1.1",
+    "prettier": "^2.8.4",
     "react-test-renderer": "^17.0.2",
     "typescript": "~4.3.5"
   },
